+++ android/src/main/java/com/revenuecat/purchases/react/RNPurchasesModule.java
@@ -43,7 +43,7 @@ public class RNPurchasesModule extends ReactContextBaseJavaModule implements Upd
 
     private static final String PURCHASER_INFO_UPDATED = "Purchases-PurchaserInfoUpdated";
     public static final String PLATFORM_NAME = "react-native";
-    public static final String PLUGIN_VERSION = "4.3.3";
+    public static final String PLUGIN_VERSION = "4.4.0";
 
     private final ReactApplicationContext reactContext;
 
@@ -294,6 +294,11 @@ public class RNPurchasesModule extends ReactContextBaseJavaModule implements Upd
     @ReactMethod
     public void setOnesignalID(String onesignalID) {
       SubscriberAttributesKt.setOnesignalID(onesignalID);
+    }
+
+    @ReactMethod
+    public void setAirshipChannelID(String airshipChannelID) {
+      SubscriberAttributesKt.setAirshipChannelID(airshipChannelID);
     }
 
     // endregion