+++ scripts/build.js
@@ -5,11 +5,11 @@ if (os.type() === "Linux") {
   console.log("Skipping iOS Dependencies");
 } else if (os.type() === "Darwin") {
   const downloadProcess = exec(
-    "./scripts/download-purchases-framework.sh 3.12.8"
+    "./scripts/download-purchases-framework.sh 3.13.0"
   );
   downloadProcess.stdout.pipe(process.stdout);
   const downloadProcessCommon = exec(
-    "./scripts/download-purchases-common.sh 1.9.3"
+    "./scripts/download-purchases-common.sh 1.10.0"
   );
   downloadProcessCommon.stdout.pipe(process.stdout);
 } else if (os.type() === "Windows_NT") {