+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 4.4.0
+
+- Bump`purchases-hybrid-common` to `1.10.0` [Changelog here](https://github.com/RevenueCat/purchases-hybrid-common/releases/tag/1.10.0)
+- Bump `purchases-ios` to `3.13.0` ([Changelog here](https://github.com/RevenueCat/purchases-ios/releases/3.13.0))
+- Bump `purchases-android` to `4.4.0` ([Changelog here](https://github.com/RevenueCat/purchases-android/releases/4.4.0))
+- Added support for Airship integration via `setAirshipChannelID`
+     https://github.com/RevenueCat/react-native-purchases/pull/307
+
 ## 4.3.3
 
 - Bump `purchases-ios` to 3.12.8